[14.0.0] Bindgen multi-version support (#7212)
* Support multiple versions in `bindgen!` (#7172)

This commit fixes a bug in the `bindgen!` macro where when faced with
multiple packages that differ only in version number invalid bindings
were generated. The fix here is to add version number information to
package module names if necessary in situations such as this. This
required some refactoring internally to have a single source of truth
for what the name of a module should be and avoid having it implicitly
calculated in two locations.

* Ensure exports are properly namespaced (#7196)

Continuation of #7172 where imports and export _definitions_ are
properly namespaced by version by the usage of exports was not.

---------

Co-authored-by: Alex Crichton <alex@alexcrichton.com>