Introduce API for custom stack memory (#7209)
* Introduce StackMemory and StackMemoryCreator

This allows custom implementations of stack memory to be plugged into
the async functionality for wasmtime. Currently, stacks are always
mmapped, and this custom allocator allows embedders to use any memory
they would like.

The new APIs are also exposed in the C api.

This has no effect on windows, as our hands are tied with fibers there.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

* Add test for custom host memory

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

* fix allocator test

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

* Address review comments

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

* Fix lint warnings

prtest:full

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

* fix windows lint warning

prtest:full

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>

---------

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>