riscv64: Improve codegen for `icmp` (#7203)
* riscv64: Constants are always sign-extended

Skip generating extra sign-extension instructions in this case because
the materialization of a constant will implicitly sign-extend into the
entire register.

* riscv64: Rename `lower_int_compare` helper

Try to reserve `lower_*` as taking a thing and producing a `*Reg`.
Rename this helper to `is_nonzero_cmp` to represent how it's testing for
nonzero and producing a comparison.

* riscv64: Rename some float comparison helpers

* `FCmp` -> `FloatCompare`
* `emit_fcmp` -> `fcmp_to_float_compare`
* `lower_fcmp` -> `lower_float_compare`

Make some room for upcoming integer comparison functions.

* riscv64: Remove `ordered` helper

This is only used by one lowering so inline its definition directly.

* riscv64: Remove the `Icmp` pseudo-instruction

This commit is the first in a few steps to reimplement #6112 and #7113.
The `Icmp` pseudo-instruction is removed here and necessary
functionality is all pushed into ISLE lowerings. This enables deleting
the `lower_br_icmp` helper in `emit.rs` as it's no longer necessary,
meaning that all conditional branches should ideally be generated in
lowering rather than pseudo-instructions to benefit from various
optimizations.

Currently the lowering is the bare-bones minimum to get things working.
This involved creating a helper to lower an `IntegerCompare` into an
`XReg` via negation/swapping args/etc. In generated code this removes
branches and makes `icmp` a straight-line instruction for non-128-bit
arguments.

* riscv64: Remove an unused helper

* riscv64: Optimize comparisons with 0

Use the `x0` register which is always zero where possible and avoid
unnecessary `xor`s against this register.

* riscv64: Specialize `a < $imm`

* riscv64: Optimize Equal/NotEqual against constants

* riscv64: Optimize LessThan with constant argument

* Optimize `a <= $imm`

* riscv64: Optimize `a >= $imm`

* riscv64: Add comment for new helper

* Use i64 in icmp optimizations

Matches the sign-extension that happens at the hardware layer.

* Correct some sign extensions

* riscv64: Don't assume immediates are extended

* riscv64: Fix encoding for `c.addi4spn`

* riscv64: Remove `icmp` lowerings which modify constants

These aren't correct and will need updating

* Add regression test

* riscv64: Fix handling unsigned comparisons with constants

---------

Co-authored-by: Afonso Bordado <afonsobordado@az8.co>