Fix profiling the first function in a module
This commit removes a clause in `register_module` which was preventing
`perf` profiling the first function in a module because it was located
at address 0. This check is relatively dated at this point and I think I
originally added it in error (and/or it was copied from somewhere else),
so this commit removes it.