Put versions in all WASI WIT files (#7272)
* Put versions in all WASI WIT files

This commit starts exercising the versioning feature of WIT by ensuring
that all WASI descriptions have a version associated with them. The
version chosen is 0.2.0 which reflects the upcoming "preview 2" release
where in theory 0.1.0 was claimed by preview1. This is intended to stay
as 0.2.0 for now and we'll determine how best to update these numbers in
the future once preview2 is released.

Closes #7171

* Allow omitting versions in `with` keys

As a convenience for now this enables omitting the version of an
interface from a `with` key. This has a risk of not working well if two
packages are present and one has a version and one doesn't, but that's
left as a PR to fix in the future as the benefit of avoiding repetition
seems good for now.

* Allow omitting versions in trappable_error_types

* Use 0.2.0-rc-2023-10-18 as a version number

* More test fixes

* Fix another test