Gate many CLI/Wasmtime features behind Cargo features (#7282)
* Move `wasmtime explore` behind a Cargo feature

Enable this Cargo feature by default, but enable building the CLI
without the `explore` subcommand.

* Move the `wast` subcommand behind a Cargo feature

* Move support for `wat` behind a CLI feature

This was already conditional in crates such as `wasmtime` and this makes
it an optional dependency of the CLI as well.

* Move CLI cache support behind a Cargo feature

Additionally refactor `wasmtime-cli-flags` to not unconditionally pull
in caching support by removing its `default` feature and appropriately
enabling it from the CLI.

* Move `rayon` behind an optional feature

* Move `http-body-util` dependency behind `serve` feature

* Add a Cargo feature for compiling out log statements

This sets the static features of `log` and `tracing` to statically
remove all log statements from the binary to cut down on binary size.

* Move logging support behind a Cargo feature

Enables statically removing logging support in addition to the previous
compiling out log statements themselves.

* Move demangling support behind a Cargo feature

* Enable building the CLI without cranelift

Compile out the `compile` subcommand for example.

* Gate all profiling support behind one feature flag

This commit moves vtune/jitdump support behind a single `profiling`
feature flag that additionally includes the guest profiler dependencies
now too.

* Move support for core dumps behind a feature flag

* Move addr2line behind a feature

* Fix rebase

* Document cargo features and a minimal build

* Tidy up the source a bit

* Rename compile-out-logging

* Document disabling logging

* Note the host architecture as well

* Fix tests

* Fix tests

* Mention debuginfo stripping

* Fix CI configuration for checking features

* Fix book tests

* Update lock file after rebase

* Enable coredump feature by default