riscv64: Refactor float-to-int trapping conversions
This commit refactors the `fcvt_to_{s,u}int` conversions in the riscv64
backend to build on the work in #7327. These two instructions currently
have multiple branches in them to check for traps when the float is out
of bounds. This is additionally signaled via the `fflags` CSR however so
this is an attempt to trim some branches/constants/etc and instead check
the `fflags` CSR after the operation. This requires clearing the CSR
just beforehand, however. This additionally doesn't work for 8/16-bit
types because the hardware doesn't set flags for those.