Remove the leftover usage of `posix-signals-on-macos` (#7360) (#7362)
This follows up #6807 and removes the last remaining reference to
the removed `posix-signals-on-macos` feature flag.

Note that `lib.rs` now imports `mod unix` on MacOS. This change
is similar to the change in `traphandlers.rs` in #6807. It is
needed for hosts that use signals instead of Mach ports on MacOs.