[release-13.0] Backport removal of `posix-signals-on-macos` (#7361)
* Remove the leftover usage of `posix-signals-on-macos` (#7360)

This follows up #6807 and removes the last remaining reference to
the removed `posix-signals-on-macos` feature flag.

Note that `lib.rs` now imports `mod unix` on MacOS. This change
is similar to the change in `traphandlers.rs` in #6807. It is
needed for hosts that use signals instead of Mach ports on MacOs.

* wasi-nn: reenable CI task (#7164)

The task was disabled due to a failing checksum. Since then, the
`install-openvino-action` has improved and now _almost_ has
cross-platform support. This change just restores the status quo in
expectation that #6895 will significantly improve the testing story.

prtest:full

---------

Co-authored-by: Andrew Brown <andrew.brown@intel.com>