[13.0.0] Add audit metadata for wasmtime-wmemcheck (#7372)
* Add audit metadata for wasmtime-wmemcheck

Required to get the release automation working again

* Fix a flaky preview2 test (#7138)

Found in CI for #7099 this commit updates the durations used in the
`resolves_immediately` and `never_resolves` tests. For immediately
resolving it should be ok to wait for a generous amount of time since
the timeout is only a protection against tests hanging. For
`never_resolves` it should be ok to wait a short amount of time and any
failure there will show up as a spurious failure.