Enable the `wasmtime serve` command by default
This enables the `serve` subcommand by default on the CLI instead of
having it behind an off-by-default feature. In general our default build
is the "most featureful" while we still retain the ability to turn off
everything if desired for custom builds. This additionally matches the
current `main` branch where through other refactorings the `serve`
command is now enabled by default.